fix(tracing,money): align suite spans and brand constructors
Thread the concrete suite span context through the shared Bun test harness so package-local tracer wrappers stop drifting, and replace the remaining money branded-constructor casts with truthful constructor-backed helpers.

+++ packages/money/src/assertions.ts
@@ -14,7 +14,7 @@ import type { Amount } from './types.js';
  *   including the imbalance value (invariant -- programmer/logic bug)
  */
 export function assertBalanced<C extends string>(legs: readonly Amount<C>[]): void {
-  const sum = legs.reduce((acc, leg) => acc + (leg as bigint), 0n);
+  const sum = legs.reduce((acc, leg) => acc + leg, 0n);
   if (sum !== 0n) {
     throw new Error(`Journal legs are not balanced: imbalance of ${sum}`);
   }
@@ -24,6 +24,6 @@ export function assertBalanced<C extends string>(legs: readonly Amount<C>[]): vo
  * Non-throwing variant -- returns true if legs sum to 0n.
  */
 export function isBalanced<C extends string>(legs: readonly Amount<C>[]): boolean {
-  const sum = legs.reduce((acc, leg) => acc + (leg as bigint), 0n);
+  const sum = legs.reduce((acc, leg) => acc + leg, 0n);
   return sum === 0n;
 }
